rose bush: improve tasks and jobs filtering
Cycles list now displays number of tasks and jobs.

Cycles list now provides download links to those cycles with a
`log/job-CYCLE.tar.gz`.

Jobs list can now be filtered by job status combinations and individual
cylc task statuses.

Rose Bush related logic from `rose.suite_engine_procs.cylc` is moved to
`rose.bush_dao`.
