+++ lib/html/static/js/rose-bush.js
@@ -155,4 +155,8 @@ $(function() {
     $("#per_page_max").click(function() {
         $("#per_page").prop("disabled", $(this).prop("checked"));
     });
+    $("#all_task_statuses").change(function() {
+        $("input.task_status").prop("disabled", $(this).prop("checked"))
+    });
+    $("#all_task_statuses").change();
 });
