+++ lib/html/template/rose-bush/cycles.html
@@ -21,7 +21,7 @@
             <input type="hidden" name="no_fuzzy_time" value="{{no_fuzzy_time}}" />
 
             <div class="row">
-              <div class="col-md-3">
+              <div class="form-group col-sm-12 col-md-6 col-lg-4">
                 <div class="input-group">
                   <label for="per_page">Entries/Page (Default=100)</label>
                   <input id="per_page" type="text" name="per_page"
@@ -34,7 +34,7 @@
                 </div>
               </div>
 
-              <div class="col-md-2">
+              <div class="form-group col-sm-12 col-md-6 col-lg-4">
                 <div class="input-group">
                   <label for="order">Sort Order</label>
                   <select name="order" title="Sort Order" class="form-control">
@@ -48,7 +48,7 @@
                 </div>
               </div>
 
-              <div class="col-md-7">
+              <div class="form-group col-sm-12 col-md-12 col-lg-4">
                 <div class="btn-group pull-right">
                   <input type="reset" class="btn btn-default" value="reset"
                   title="Reset Display Options"/>
@@ -82,31 +82,89 @@
 </div>
 
 <div class="container-fluid"><!-- entries -->
-  {% for entry in entries -%}
-    <div class="row entry">
-      <div class="col-md-2">
-        <a href="{{script}}/jobs/{{user}}/{{suite}}?cycles={{entry.cycle|replace('+', '%2B')}}" class="cycle" title="{{entry.cycle}}">{{entry.cycle}}</a>
-      </div>
-      {% for state, icon, label, title in [("active", "play", "info", "active tasks"), ("success", "ok", "success", "succeeded tasks"), ("fail", "warning-sign", "danger", "failed tasks")] -%}
-        {% set n_state = entry.n_states[state] -%}
-        <div class="col-md-2">
-          {% if n_state -%}
-            {% set state_url = script ~ "/jobs/" ~ user ~ "/" ~ suite ~ "?cycles=" ~ entry.cycle|replace('+', '%2B') -%}
-            <a href="{{state_url}}{% for no_state in ["active", "success", "fail"] if no_state != state %}&amp;no_status={{no_state}}{% endfor %}">
-          {% endif -%}
-          <span class="label{% if n_state %} label-{{label}}{% else %} label-default{% endif %}"><i class="glyphicon glyphicon-{{icon}}" title="{{state}}"></i></span>
-          <span class="label badge-label {% if n_state %}label-{{label}}{% else -%}label-default{% endif %}" title="{{title}}">{{entry.n_states[state]}}</span>
-          {% if n_state %}</a>{% endif %}
-          {% if state=="fail" and entry.n_states["job_fails"] > 0 %}(<span class="badge  badge-{{label}}" title="job failures">{{entry.n_states["job_fails"]}}</span>){% endif %}
-      </div>
-      {% endfor -%}
+{% for entry in entries -%}
+  {% set cycle_in_url = entry.cycle|replace('+', '%2B') -%}
+  {% set task_jobs_url = (
+      script ~ "/taskjobs/" ~ user ~ "/" ~ suite ~ "?cycles=" ~ cycle_in_url
+  ) -%}
+  <div class="row entry">
 
-      <div class="col-md-4">
-        <span class="label label-default"><i class="glyphicon glyphicon-time" title="last active"></i> last active</span>
-        <abbr class="livestamp" title="{{entry.max_time_updated}}">{{entry.max_time_updated}}</abbr>
-      </div>
-
+    <div class="col-xs-6 col-md-3">
+      <a href="{{task_jobs_url}}"
+      class="cycle" title="{{entry.cycle}}">{{entry.cycle}}</a>
     </div>
-  {% endfor -%}
+
+    <div class="col-xs-6 col-md-3">
+      last active
+      <abbr class="livestamp"
+      title="{{entry.max_time_updated}}">{{entry.max_time_updated}}</abbr>
+    </div>
+
+  {# task and jobs states -#}
+  {% for state, icon, label, title_, unit1, unit2, url_arg in [
+      ("active", "play", "info", "active", "task", "tasks",
+       "&amp;task_status=" ~
+       task_status_groups["active"]|join("&amp;task_status=")),
+      ("job_active", "play-circle", "info", "active", "job", "jobs",
+       "&amp;job_status=submitted,running"),
+      ("success", "ok", "success", "succeeded", "task", "tasks",
+       "&amp;task_status=" ~
+       task_status_groups["success"]|join("&amp;task_status=")),
+      ("job_success", "ok-circle", "success", "succeeded", "job", "jobs",
+       "&amp;job_status=succeeded"),
+      ("fail", "remove", "danger", "failed", "task", "tasks",
+       "&amp;task_status=" ~
+       task_status_groups["fail"]|join("&amp;task_status=")),
+      ("job_fail", "remove-circle", "danger", "failed", "job", "jobs",
+       "&amp;job_status=submission-failed,failed"),
+  ] -%}
+    {% set n_state = entry.n_states[state] -%}
+    {% set unit = unit1 -%}
+    {% if n_state -%}
+      {% set more_label = "label-" ~ label -%}
+      {% if n_state > 1 -%}
+        {% set unit = unit2 -%}
+      {% endif -%}
+    {% else -%}
+      {% set more_label = "label-default" -%}
+    {% endif -%}
+
+    <div class="col-xs-2 col-md-1">
+    {% if n_state -%}
+      <a href="{{task_jobs_url}}{{url_arg}}">
+        <span class="label {{more_label}}"
+        title="{{n_state}} {{title_}} {{unit}}">
+          <i class="glyphicon glyphicon-{{icon}}"></i>
+          {{n_state}}
+        </span>
+      </a>
+    {% else -%}
+      <span class="label {{more_label}}"
+      title="{{n_state}} {{title_}} {{unit}}">
+        <i class="glyphicon glyphicon-{{icon}}"></i>
+        {{n_state}}
+      </span>
+    {% endif -%}
+    </div>
+
+  {% endfor -%}{# task and jobs states -#}
+
+  {% if entry.has_log_job_tar_gz -%}
+    {% set download_text = "path=log/job-" ~ entry.cycle ~ ".tar.gz" -%}
+    {% set download = (
+        script ~ "/view/" ~ user ~ "/" ~ suite ~ "?path=log/job-" ~
+        cycle_in_url ~ ".tar.gz&amp;mode=download"
+    ) -%}
+    <div class="col-xs-12 col-md-12">
+      <small>
+        <a href="{{download}}"
+        download="{{user}}-{{suite}}-log-job-{{entry.cycle}}.tar.gz">download
+        {{download_text}}</a>
+      </small>
+    </div>
+  {% endif -%}
+
+  </div>
+{% endfor -%}{# for entry in entries -#}
 </div>
 {% endblock %}
