+++ lib/html/template/rose-bush/index.html
@@ -22,7 +22,7 @@ rel="stylesheet" media="screen" />
 <div class="row">
 <div class="col-md-4">
 <form action="{{script}}/suites" class="form-inline">
-<fieldset>
+<fieldset class="form-group">
 <legend>User's Suites List:</legend>
 <input id="user" name="user" type="text" class="form-control"
 placeholder="User ID" required="1" />
@@ -31,8 +31,8 @@ placeholder="User ID" required="1" />
 </form>
 </div>
 <div class="col-md-8">
-<form action="{{script}}/jobs" class="form-inline">
-<fieldset>
+<form action="{{script}}/taskjobs" class="form-inline">
+<fieldset class="form-group">
 <legend>User's Suite Cycles or Jobs List:</legend>
 <input id="user" name="user" type="text" class="input-short form-control"
 placeholder="User ID" required="1" />
