+++ lib/html/template/rose-bush/suite-state.html
@@ -2,7 +2,7 @@
   Suite
   {% if states.is_failed -%}
     <span class="label label-danger"><i
-    class="glyphicon glyphicon-warning-sign"
+    class="glyphicon glyphicon-remove"
     title="failed task(s)"></i></span>
     has failed tasks,
   {% endif -%}
