+++ lib/html/template/rose-bush/suites.html
@@ -148,15 +148,16 @@ rel="stylesheet" media="screen" />
 {{entry.name}}
 {% endif -%}
 </div>
-<div class="col-md-2">
+<div class="col-md-1">
   <a href="{{script}}/cycles/{{user}}/{{entry.name}}">
     <button type="button" class="btn btn-default btn-sm">
       <span class="glyphicon glyphicon-tasks" aria-hidden="true"></span>
       cycles list
     </button>
   </a>
-
-  <a href="{{script}}/jobs/{{user}}/{{entry.name}}">
+</div>
+<div class="col-md-1">
+  <a href="{{script}}/taskjobs/{{user}}/{{entry.name}}">
     <button type="button" class="btn btn-default btn-sm">
       <span class="glyphicon glyphicon-list" aria-hidden="true"></span>
       jobs list
