+++ lib/python/rose/suite_engine_procs/cylc.py
@@ -20,7 +20,6 @@
 """Logic specific to the Cylc suite engine."""
 
 import filecmp
-from fnmatch import fnmatch
 from glob import glob
 import os
 from pipes import quote
@@ -40,7 +39,7 @@ import socket
 import sqlite3
 import tarfile
 from tempfile import mkstemp
-from time import sleep, time, gmtime, strftime
+from time import sleep, time
 from uuid import uuid4
 
 
@@ -52,99 +51,20 @@ class CylcProcessor(SuiteEngineProcessor):
 
     """Logic specific to the cylc suite engine."""
 
-    CYCLE_ORDERS = {"time_desc": " DESC", "time_asc": " ASC"}
-    EVENTS = {"submission succeeded": "submit",
-              "submission failed": "fail(submit)",
-              "submitting now": "submit-init",
-              "incrementing submit number": "submit-init",
-              "started": "init",
-              "succeeded": "success",
-              "failed": "fail",
-              "execution started": "init",
-              "execution succeeded": "success",
-              "execution failed": "fail",
-              "signaled": "fail(%s)"}
-    EVENT_TIME_INDICES = {
-        "submit-init": 0, "init": 1, "success": 2, "fail": 2, "fail(%s)": 2}
-    EVENT_RANKS = {"submit-init": 0, "submit": 1, "fail(submit)": 1, "init": 2,
-                   "success": 3, "fail": 3, "fail(%s)": 4}
-    JOB_LOGS_DB = "log/rose-job-logs.db"
-    JOB_ORDERS_OLD = {
-        "time_desc": "time DESC, submit_num DESC, name DESC, cycle DESC",
-        "time_asc": "time ASC, submit_num ASC, name ASC, cycle ASC",
-        "cycle_desc_name_asc": "cycle DESC, name ASC, submit_num DESC",
-        "cycle_desc_name_desc": "cycle DESC, name DESC, submit_num DESC",
-        "cycle_asc_name_asc": "cycle ASC, name ASC, submit_num DESC",
-        "cycle_asc_name_desc": "cycle ASC, name DESC, submit_num DESC",
-        "name_asc_cycle_asc": "name ASC, cycle ASC, submit_num DESC",
-        "name_desc_cycle_asc": "name DESC, cycle ASC, submit_num DESC",
-        "name_asc_cycle_desc": "name ASC, cycle DESC, submit_num DESC",
-        "name_desc_cycle_desc": "name DESC, cycle DESC, submit_num DESC"}
-    JOB_ORDERS = dict(JOB_ORDERS_OLD)
-    JOB_ORDERS.update({
-        "time_submit_desc": (
-            "time_submit DESC, submit_num DESC, name DESC, cycle DESC"),
-        "time_submit_asc": (
-            "time_submit ASC, submit_num DESC, name DESC, cycle DESC"),
-        "time_run_desc": (
-            "time_run DESC, submit_num DESC, name DESC, cycle DESC"),
-        "time_run_asc": (
-            "time_run ASC, submit_num DESC, name DESC, cycle DESC"),
-        "time_run_exit_desc": (
-            "time_run_exit DESC, submit_num DESC, name DESC, cycle DESC"),
-        "time_run_exit_asc": (
-            "time_run_exit ASC, submit_num DESC, name DESC, cycle DESC"),
-        "duration_queue_desc": (
-            "(CAST(strftime('%s', time_run) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_submit) AS NUMERIC)) DESC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-        "duration_queue_asc": (
-            "(CAST(strftime('%s', time_run) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_submit) AS NUMERIC)) ASC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-        "duration_run_desc": (
-            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_run) AS NUMERIC)) DESC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-        "duration_run_asc": (
-            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_run) AS NUMERIC)) ASC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-        "duration_queue_run_desc": (
-            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_submit) AS NUMERIC)) DESC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-        "duration_queue_run_asc": (
-            "(CAST(strftime('%s', time_run_exit) AS NUMERIC) -" +
-            " CAST(strftime('%s', time_submit) AS NUMERIC)) ASC, " +
-            "submit_num DESC, name DESC, cycle DESC"),
-    })
     PGREP_CYLC_RUN = r"python.*/bin/cylc-(run|restart)( | .+ )%s( |$)"
-    REASON_KEY_PROC = "process"
-    REASON_KEY_FILE = "port-file"
     REC_CYCLE_TIME = re.compile(
         r"\A[\+\-]?\d+(?:W\d+)?(?:T\d+(?:Z|[+-]\d+)?)?\Z")  # Good enough?
-    REC_SEQ_LOG = re.compile(r"\A(.+\.)([^\.]+)(\.[^\.]+)\Z")
-    REC_SIGNALLED = re.compile(r"Task\sjob\sscript\sreceived\ssignal\s(\S+)")
     SCHEME = "cylc"
-    STATUSES = {"active": ["ready", "queued", "submitting", "submitted",
-                           "submit-retrying", "running", "retrying"],
-                "fail": ["submission failed", "failed"],
-                "success": ["expired", "succeeded"]}
     SUITE_CONF = "suite.rc"
     SUITE_DB = "cylc-suite.db"
     SUITE_DIR_REL_ROOT = "cylc-run"
     TASK_ID_DELIM = "."
+
     TIMEOUT = 60  # seconds
 
     def __init__(self, *args, **kwargs):
         SuiteEngineProcessor.__init__(self, *args, **kwargs)
-        self.daos = {self.SUITE_DB: {}, self.JOB_LOGS_DB: {}}
-        # N.B. Should be considered a constant after initialisation
-        self.state_of = {}
-        for status, names in self.STATUSES.items():
-            for name in names:
-                self.state_of[name] = status
+        self.daos = {}
         self.host = None
         self.user = None
 
@@ -231,474 +151,21 @@ class CylcProcessor(SuiteEngineProcessor):
         self.popen(fmt % (host, suite_name, args_str, os.devnull),
                    env=environ, shell=True)
 
-    def get_suite_broadcast_states(self, user_name, suite_name):
-        """Return broadcast states of a suite.
-
-        [[point, name, key, value], ...]
-
-        """
-        # Check if "broadcast_states" table is available or not
-        if not self._db_has_table(
-                self.SUITE_DB, user_name, suite_name, "broadcast_states"):
-            return
-
-        broadcast_states = []
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name,
-                "SELECT point,namespace,key,value FROM broadcast_states" +
-                " ORDER BY point ASC, namespace ASC, key ASC"):
-            point, namespace, key, value = row
-            broadcast_states.append([point, namespace, key, value])
-        return broadcast_states
-
-    def get_suite_broadcast_events(self, user_name, suite_name):
-        """Return broadcast events of a suite.
-
-        [[time, change, point, name, key, value], ...]
-
-        """
-        # Check if "broadcast_events" table is available or not
-        if not self._db_has_table(
-                self.SUITE_DB, user_name, suite_name, "broadcast_events"):
-            return {}
-
-        broadcast_events = []
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name,
-                "SELECT time,change,point,namespace,key,value" +
-                " FROM broadcast_events" +
-                " ORDER BY time DESC, point DESC, namespace DESC, key DESC"):
-            time, change, point, namespace, key, value = row
-            broadcast_events.append(
-                (time, change, point, namespace, key, value))
-        return broadcast_events
-
-    def get_suite_dir_rel(self, suite_name, *paths):
+    @classmethod
+    def get_suite_dir_rel(cls, suite_name, *paths):
         """Return the relative path to the suite running directory.
 
         paths -- if specified, are added to the end of the path.
 
         """
-        return os.path.join(self.SUITE_DIR_REL_ROOT, suite_name, *paths)
-
-    def get_suite_job_events(
-            self, user_name, suite_name, cycles, tasks, no_statuses, order,
-            limit, offset):
-        """Return suite task job events.
-
-        user -- A string containing a valid user ID
-        suite -- A string containing a valid suite ID
-        cycles -- Display only task jobs matching these cycles. A value in the
-                  list can be a cycle, the string "before|after CYCLE", or a
-                  glob to match cycles.
-        tasks -- Display only jobs for task names matching these names. Values
-                 can be a valid task name or a glob like pattern for matching
-                 valid task names.
-        no_statuses -- Do not display jobs with these statuses. Valid values
-                       are the keys of CylcProcessor.STATUSES.
-        order -- Order search in a predetermined way. A valid value is one of
-                 the keys in CylcProcessor.ORDERS.
-        limit -- Limit number of returned entries
-        offset -- Offset entry number
-
-        Return (entries, of_n_entries) where:
-        entries -- A list of matching entries
-        of_n_entries -- Total number of entries matching query
-
-        Each entry is a dict:
-            {"cycle": cycle, "name": name, "submit_num": submit_num,
-             "events": [time_submit, time_init, time_exit],
-             "status": None|"submit|fail(submit)|init|success|fail|fail(%s)",
-             "logs": {"script": {"path": path, "path_in_tar", path_in_tar,
-                                 "size": size, "mtime": mtime},
-                      "out": {...},
-                      "err": {...},
-                      ...}}
-
-        """
-        # Check if "task_jobs" table is available or not
-        if not self._db_has_table(
-                self.SUITE_DB, user_name, suite_name, "task_jobs"):
-            return self._get_suite_job_events_old(
-                user_name, suite_name, cycles, tasks, no_statuses, order,
-                limit, offset)
-
-        # Build the WHERE expression to filter by cycles, tasks, statuses
-        where_exprs = []
-        where_args = []
-        if cycles:
-            cycle_where_exprs = []
-            for cycle in cycles:
-                if cycle.startswith("before "):
-                    where_args.append(cycle.split(None, 1)[-1])
-                    cycle_where_exprs.append("cycle <= ?")
-                elif cycle.startswith("after "):
-                    where_args.append(cycle.split(None, 1)[-1])
-                    cycle_where_exprs.append("cycle >= ?")
-                else:
-                    where_args.append(cycle)
-                    cycle_where_exprs.append("cycle GLOB ?")
-            where_exprs.append(" OR ".join(cycle_where_exprs))
-        if tasks:
-            where_exprs.append(" OR ".join(["name GLOB ?"] * len(tasks)))
-            where_args += tasks
-        if no_statuses:
-            for no_status in no_statuses:
-                statuses = self.STATUSES.get(no_status, [])
-                where_exprs.append(
-                    " AND ".join(["status != ?"] * len(statuses)))
-                where_args += statuses
-        if where_exprs:
-            where_expr = " WHERE (" + ") AND (".join(where_exprs) + ")"
-        else:
-            where_expr = ""
-
-        # Get number of entries
-        of_n_entries = 0
-        stmt = ("SELECT COUNT(*)" +
-                " FROM task_jobs JOIN task_states USING (name, cycle)" +
-                where_expr)
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt, where_args):
-            of_n_entries = row[0]
-            break
-        else:
-            self._db_close(self.SUITE_DB, user_name, suite_name)
-            return ([], 0)
-
-        # Get entries
-        entries = []
-        entry_of = {}
-        stmt = ("SELECT" +
-                " task_states.time_updated AS time," +
-                " cycle, name," +
-                " task_jobs.submit_num AS submit_num," +
-                " task_states.submit_num AS submit_num_max," +
-                " task_states.status AS task_status," +
-                " time_submit, submit_status," +
-                " time_run, time_run_exit, run_signal, run_status," +
-                " user_at_host, batch_sys_name, batch_sys_job_id" +
-                " FROM task_jobs JOIN task_states USING (cycle, name)" +
-                where_expr +
-                " ORDER BY " +
-                self.JOB_ORDERS.get(order, self.JOB_ORDERS["time_desc"]))
-        limit_args = []
-        if limit:
-            stmt += " LIMIT ? OFFSET ?"
-            limit_args = [limit, offset]
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt,
-                where_args + limit_args):
-            (
-                cycle, name, submit_num, submit_num_max, task_status,
-                time_submit, submit_status,
-                time_run, time_run_exit, run_signal, run_status,
-                user_at_host, batch_sys_name, batch_sys_job_id
-            ) = row[1:]
-            entry = {
-                "cycle": cycle,
-                "name": name,
-                "submit_num": submit_num,
-                "submit_num_max": submit_num_max,
-                "events": [time_submit, time_run, time_run_exit],
-                "task_status": None,
-                "status": None,
-                "host": user_at_host,
-                "submit_method": batch_sys_name,
-                "submit_method_id": batch_sys_job_id,
-                "logs": {},
-                "seq_logs_indexes": {}}
-            try:
-                entry["task_status"] = self.state_of[task_status]
-            except KeyError:
-                pass
-            if run_status and run_signal:
-                entry["status"] = "fail(%s)" % (run_signal)
-            elif run_status:
-                entry["status"] = "fail"
-            elif run_status == 0:
-                entry["status"] = "success"
-            elif submit_status:
-                entry["status"] = "fail(submit)"
-            elif submit_status == 0 and time_run:
-                entry["status"] = "init"
-            elif submit_status == 0:
-                entry["status"] = "submit"
-            else:
-                entry["status"] = "submit-init"
-            entries.append(entry)
-            entry_of[(cycle, name, submit_num)] = entry
-        self._db_close(self.SUITE_DB, user_name, suite_name)
-        if not entries:
-            return (entries, of_n_entries)
-
-        # Job logs
-        # Recent job logs are likely to be in the file system, so we can get a
-        # listing of the relevant "log/job/CYCLE/NAME/SUBMI_NUM/" directory.
-        # Older job logs may be archived in "log/job-CYCLE.tar.gz", we should
-        # only open each relevant TAR file once to obtain a listing for all
-        # relevant entries of that cycle.
-        prefix = "~"
-        if user_name:
-            prefix += user_name
-        user_suite_dir = os.path.expanduser(os.path.join(
-            prefix, self.get_suite_dir_rel(suite_name)))
-        try:
-            fs_log_cycles = os.listdir(
-                os.path.join(user_suite_dir, "log", "job"))
-        except OSError:
-            fs_log_cycles = []
-        targzip_log_cycles = []
-        for name in glob(os.path.join(user_suite_dir, "log", "job-*.tar.gz")):
-            targzip_log_cycles.append(os.path.basename(name)[4:-7])
-
-        relevant_targzip_log_cycles = []
-        for entry in entries:
-            if entry["cycle"] in fs_log_cycles:
-                pathd = "log/job/%(cycle)s/%(name)s/%(submit_num)02d" % entry
-                try:
-                    filenames = os.listdir(os.path.join(user_suite_dir, pathd))
-                except OSError:
-                    pass
-                for filename in filenames:
-                    try:
-                        stat = os.stat(
-                            os.path.join(user_suite_dir, pathd, filename))
-                    except OSError:
-                        pass
-                    entry["logs"][filename] = {
-                        "path": "/".join([pathd, filename]),
-                        "path_in_tar": None,
-                        "mtime": int(stat.st_mtime),  # too precise otherwise
-                        "size": stat.st_size,
-                        "exists": True,
-                        "seq_key": None}
-            elif entry["cycle"] in targzip_log_cycles:
-                if entry["cycle"] not in relevant_targzip_log_cycles:
-                    relevant_targzip_log_cycles.append(entry["cycle"])
-
-        for cycle in relevant_targzip_log_cycles:
-            path = os.path.join("log", "job-%s.tar.gz" % cycle)
-            tar = tarfile.open(os.path.join(user_suite_dir, path), "r:gz")
-            for member in tar.getmembers():
-                # member.name expected to be "job/cycle/task/submit_num/*"
-                if not member.isfile():
-                    continue
-                try:
-                    cycle_str, name, submit_num_str = (
-                        member.name.split("/", 4)[1:4])
-                    entry = entry_of[(cycle_str, name, int(submit_num_str))]
-                except (KeyError, ValueError):
-                    continue
-                entry["logs"][os.path.basename(member.name)] = {
-                    "path": path,
-                    "path_in_tar": member.name,
-                    "mtime": int(member.mtime),  # too precise otherwise
-                    "size": member.size,
-                    "exists": True,
-                    "seq_key": None}
-
-        # Sequential logs
-        for entry in entries:
-            for filename, filename_items in entry["logs"].items():
-                seq_log_match = self.REC_SEQ_LOG.match(filename)
-                if not seq_log_match:
-                    continue
-                head, index_str, tail = seq_log_match.groups()
-                seq_key = head + "*" + tail
-                filename_items["seq_key"] = seq_key
-                if seq_key not in entry["seq_logs_indexes"]:
-                    entry["seq_logs_indexes"][seq_key] = {}
-                entry["seq_logs_indexes"][seq_key][index_str] = filename
-            for seq_key, indexes in entry["seq_logs_indexes"].items():
-                # Only one item, not a sequence
-                if len(indexes) <= 1:
-                    entry["seq_logs_indexes"].pop(seq_key)
-                # All index_str are numbers, convert key to integer so
-                # the template can sort them as numbers
-                try:
-                    int_indexes = {}
-                    for index_str, filename in indexes.items():
-                        int_indexes[int(index_str)] = filename
-                    entry["seq_logs_indexes"][seq_key] = int_indexes
-                except ValueError:
-                    pass
-            for filename, log_dict in entry["logs"].items():
-                # Unset seq_key for singular items
-                if log_dict["seq_key"] not in entry["seq_logs_indexes"]:
-                    log_dict["seq_key"] = None
-
-        return (entries, of_n_entries)
-
-    def _get_suite_job_events_old(
-            self, user_name, suite_name, cycles, tasks, no_statuses, order,
-            limit, offset):
-        """The "get_suite_job_events" method for older cylc versions."""
-        # Build WHERE expression to select by cycles and/or task names
-        where = ""
-        stmt_args = []
-        if cycles:
-            where_fragments = []
-            for cycle in cycles:
-                if cycle.startswith("before "):
-                    stmt_args.append(cycle.split(None, 1)[-1])
-                    where_fragments.append("cycle <= ?")
-                elif cycle.startswith("after "):
-                    stmt_args.append(cycle.split(None, 1)[-1])
-                    where_fragments.append("cycle >= ?")
-                else:
-                    stmt_args.append(cycle)
-                    where_fragments.append("cycle GLOB ?")
-            where += " AND (" + " OR ".join(where_fragments) + ")"
-        if tasks:
-            where_fragments = []
-            for task in tasks:
-                where_fragments.append("name GLOB ?")
-                stmt_args.append(task)
-            where += " AND (" + " OR ".join(where_fragments) + ")"
-        if no_statuses:
-            where_fragments = []
-            for no_status in no_statuses:
-                for status in self.STATUSES.get(no_status, []):
-                    where_fragments.append("status != ?")
-                    stmt_args.append(status)
-            where += " AND (" + " AND ".join(where_fragments) + ")"
-        # Execute query to get number of entries
-        of_n_entries = 0
-        stmt = ("SELECT COUNT(*) FROM"
-                " task_events JOIN task_states USING (name,cycle)"
-                " WHERE event==? OR event==?")
-        if where:
-            stmt += " " + where
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt,
-                ["submitting now", "incrementing submit number"] + stmt_args):
-            of_n_entries = row[0]
-            break
-        if not of_n_entries:
-            return ([], 0)
-        # Execute query to get entries
-        entries = []
-        stmt = (
-            "SELECT" +
-            " cycle, name, task_events.submit_num AS submit_num," +
-            " group_concat(time), group_concat(event)," +
-            " group_concat(message) " +
-            " FROM" +
-            " task_events JOIN task_states USING (cycle,name)" +
-            " WHERE" +
-            " (event==? OR event==? OR event==? OR" +
-            "  event==? OR event==? OR event==? OR event==?)" +
-            where +
-            " GROUP BY cycle, name, task_events.submit_num" +
-            " ORDER BY " +
-            self.JOB_ORDERS_OLD.get(order, self.JOB_ORDERS_OLD["time_desc"]))
-        stmt_args_head = [
-            "submitting now", "incrementing submit number",
-            "submission failed", "started", "succeeded", "failed", "signaled"]
-        stmt_args_tail = []
-        if limit:
-            stmt += " LIMIT ? OFFSET ?"
-            stmt_args_tail = [limit, offset]
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt,
-                stmt_args_head + stmt_args + stmt_args_tail):
-            cycle, name, submit_num, times_str, events_str, messages_str = row
-            entry = {"cycle": cycle, "name": name, "submit_num": submit_num,
-                     "submit_num_max": 1, "events": [None, None, None],
-                     "status": None, "logs": {}, "seq_logs_indexes": {}}
-            entries.append(entry)
-            events = events_str.split(",")
-            times = times_str.split(",")
-            event_rank = -1
-            for event, time_ in zip(events, times):
-                my_event = self.EVENTS.get(event)
-                if self.EVENT_TIME_INDICES.get(my_event) is not None:
-                    entry["events"][self.EVENT_TIME_INDICES[my_event]] = time_
-                if (self.EVENT_RANKS.get(my_event) is not None and
-                        self.EVENT_RANKS[my_event] > event_rank):
-                    entry["status"] = my_event
-                    event_rank = self.EVENT_RANKS[my_event]
-                    if my_event == "fail(%s)":
-                        match = self.REC_SIGNALLED.search(messages_str)
-                        if match:
-                            entry["status"] = "fail(%s)" % match.group(1)
-                        else:
-                            entry["status"] = "fail(SIGNAL)"
-
-        other_info_of = {}
-        for entry in entries:
-            cycle = entry["cycle"]
-            name = entry["name"]
-            if (cycle, name) not in other_info_of:
-                stmt = ("SELECT" +
-                        " submit_num,host,submit_method,submit_method_id" +
-                        " FROM task_states" +
-                        " WHERE cycle==? AND name==?")
-                for row in self._db_exec(self.SUITE_DB, user_name, suite_name,
-                                         stmt, [cycle, name]):
-                    other_info_of[(cycle, name)] = list(row)
-                    break
-            entry["submit_num_max"] = other_info_of[(cycle, name)][0]
-            if entry["submit_num"] == entry["submit_num_max"]:
-                entry["host"] = other_info_of[(cycle, name)][1]
-                entry["submit_method"] = other_info_of[(cycle, name)][2]
-                entry["submit_method_id"] = other_info_of[(cycle, name)][3]
-
-        self._db_close(self.SUITE_DB, user_name, suite_name)
-
-        # Job logs DB
-        stmt = ("SELECT key,path,path_in_tar,mtime,size FROM log_files " +
-                "WHERE cycle==? AND task==? AND submit_num==?")
-        prefix = "~"
-        if user_name:
-            prefix += user_name
-        user_suite_dir = os.path.join(os.path.expanduser(prefix),
-                                      self.get_suite_dir_rel(suite_name))
-        for entry in entries:
-            stmt_args = [entry["cycle"], entry["name"], entry["submit_num"]]
-            cursor = self._db_exec(self.JOB_LOGS_DB, user_name, suite_name,
-                                   stmt, stmt_args)
-            if cursor is None:
-                continue
-            for row in cursor:
-                key, path, path_in_tar, mtime, size = row
-                abs_path = os.path.join(user_suite_dir, path)
-                entry["logs"][key] = {"path": path,
-                                      "path_in_tar": path_in_tar,
-                                      "mtime": mtime,
-                                      "size": size,
-                                      "exists": os.path.exists(abs_path),
-                                      "seq_key": None}
-                seq_log_match = self.REC_SEQ_LOG.match(key)
-                if not seq_log_match:
-                    continue
-                head, index_str, tail = seq_log_match.groups()
-                if not tail and not index_str.isdigit():
-                    continue
-                if not tail:
-                    tail = ""
-                seq_key = head + "*" + tail
-                entry["logs"][key]["seq_key"] = seq_key
-                if seq_key not in entry["seq_logs_indexes"]:
-                    entry["seq_logs_indexes"][seq_key] = {}
-                entry["seq_logs_indexes"][seq_key][index_str] = key
-
-            for seq_key, indexes in entry["seq_logs_indexes"].items():
-                if len(indexes) <= 1:
-                    entry["seq_logs_indexes"].pop(seq_key)
-            for key, log_dict in entry["logs"].items():
-                if log_dict["seq_key"] not in entry["seq_logs_indexes"]:
-                    log_dict["seq_key"] = None
-        self._db_close(self.JOB_LOGS_DB, user_name, suite_name)
-        return (entries, of_n_entries)
+        return os.path.join(cls.SUITE_DIR_REL_ROOT, suite_name, *paths)
 
     def get_suite_jobs_auths(self, suite_name, cycle_time=None,
                              task_name=None):
         """Return remote ["[user@]host", ...] for submitted jobs."""
         auths = []
         # Check if "task_jobs" table is available or not
-        if self._db_has_table(self.SUITE_DB, None, suite_name, "task_jobs"):
+        if self._db_has_table(suite_name, "task_jobs"):
             stmt = "SELECT DISTINCT user_at_host FROM task_jobs"
             stmt_where_list = []
             stmt_args = []
@@ -712,180 +179,13 @@ class CylcProcessor(SuiteEngineProcessor):
                 stmt_args.append(arg)
         if stmt_where_list:
             stmt += " WHERE " + " AND ".join(stmt_where_list)
-        for row in self._db_exec(
-                self.SUITE_DB, None, suite_name, stmt, stmt_args):
+        for row in self._db_exec(suite_name, stmt, stmt_args):
             if row and row[0]:
                 auth = self._parse_user_host(auth=row[0])
                 if auth:
                     auths.append(auth)
-        self._db_close(self.SUITE_DB, None, suite_name)
+        self._db_close(suite_name)
         return auths
-
-    def get_suite_logs_info(self, user_name, suite_name):
-        """Return the information of the suite logs.
-
-        Return a tuple that looks like:
-            ("cylc-run",
-             {"err": {"path": "log/suite/err", "mtime": mtime, "size": size},
-              "log": {"path": "log/suite/log", "mtime": mtime, "size": size},
-              "out": {"path": "log/suite/out", "mtime": mtime, "size": size}})
-
-        """
-        logs_info = {}
-        prefix = "~"
-        if user_name:
-            prefix += user_name
-        d_rel = self.get_suite_dir_rel(suite_name)
-        dir_ = os.path.expanduser(os.path.join(prefix, d_rel))
-        # Get cylc files.
-        cylc_files = ["cylc-suite-env", "suite.rc", "suite.rc.processed"]
-        for key in cylc_files:
-            f_name = os.path.join(dir_, key)
-            if os.path.isfile(f_name):
-                f_stat = os.stat(f_name)
-                logs_info[key] = {"path": key,
-                                  "mtime": f_stat.st_mtime,
-                                  "size": f_stat.st_size}
-        # Get cylc suite log files.
-        log_files = ["log/suite/err", "log/suite/log", "log/suite/out"]
-        for key in log_files:
-            f_name = os.path.join(dir_, key)
-            if os.path.isfile(f_name):
-                try:
-                    link_path = os.readlink(f_name)
-                except OSError:
-                    link_path = f_name
-                old_logs = []  # Old log naming system.
-                new_logs = []  # New log naming system.
-                # TODO: Post migration to cylc this logic can be replaced by:
-                # `from cylc.suite_logging import get_logs` (superior)
-                for log in glob(f_name + '.*'):
-                    log_name = os.path.basename(log)
-                    if log_name == link_path:
-                        continue
-                    if len(log_name.split('.')[1]) > 3:
-                        new_logs.append(os.path.join("log", "suite", log_name))
-                    else:
-                        old_logs.append(os.path.join("log", "suite", log_name))
-                new_logs.sort(reverse=True)
-                old_logs.sort()
-                f_stat = os.stat(f_name)
-                logs_info[key] = {"path": key,
-                                  "paths": [key] + new_logs + old_logs,
-                                  "mtime": f_stat.st_mtime,
-                                  "size": f_stat.st_size}
-        return ("cylc", logs_info)
-
-    def get_suite_cycles_summary(
-            self, user_name, suite_name, order, limit, offset):
-        """Return a the state summary (of each cycle) of a user's suite.
-
-        user -- A string containing a valid user ID
-        suite -- A string containing a valid suite ID
-        limit -- Limit number of returned entries
-        offset -- Offset entry number
-
-        Return (entries, of_n_entries), where entries is a data structure that
-        looks like:
-            [   {   "cycle": cycle,
-                    "n_states": {
-                        "active": N, "success": M, "fail": L, "job_fails": K,
-                    },
-                    "max_time_updated": T2,
-                },
-                # ...
-            ]
-        where:
-        * cycle is a date-time cycle label
-        * N, M, L, K are the numbers of tasks in given states
-        * T2 is the time when last update time of (a task in) the cycle
-
-        and of_n_entries is the total number of entries.
-
-        """
-        of_n_entries = 0
-        stmt = ("SELECT COUNT(DISTINCT cycle) FROM task_states WHERE " +
-                "submit_num > 0")
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt):
-            of_n_entries = row[0]
-            break
-        if not of_n_entries:
-            return ([], 0)
-
-        # FIXME: Not strictly correct, if cycle is in basic date-only format
-        integer_mode = False
-        stmt = "SELECT cycle FROM task_states LIMIT 1"
-        for row in self._db_exec(self.SUITE_DB, user_name, suite_name, stmt):
-            integer_mode = row[0].isdigit()
-            break
-
-        states_stmt = {}
-        for key, names in self.STATUSES.items():
-            states_stmt[key] = " OR ".join(
-                ["status=='%s'" % (name) for name in names])
-        stmt = (
-            "SELECT" +
-            " cycle," +
-            " max(time_updated)," +
-            " sum(" + states_stmt["active"] + ") AS n_active," +
-            " sum(" + states_stmt["success"] + ") AS n_success,"
-            " sum(" + states_stmt["fail"] + ") AS n_fail"
-            " FROM task_states" +
-            " GROUP BY cycle")
-        if integer_mode:
-            stmt += " ORDER BY cast(cycle as number)"
-        else:
-            stmt += " ORDER BY cycle"
-        stmt += self.CYCLE_ORDERS.get(order, self.CYCLE_ORDERS["time_desc"])
-        stmt_args = []
-        if limit:
-            stmt += " LIMIT ? OFFSET ?"
-            stmt_args += [limit, offset]
-        entry_of = {}
-        entries = []
-        for row in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt, stmt_args):
-            cycle, max_time_updated, n_active, n_success, n_fail = row
-            if n_active or n_success or n_fail:
-                entry_of[cycle] = {
-                    "cycle": cycle,
-                    "max_time_updated": max_time_updated,
-                    "n_states": {
-                        "active": n_active,
-                        "success": n_success,
-                        "fail": n_fail,
-                        "job_fails": 0,
-                    },
-                }
-                entries.append(entry_of[cycle])
-
-        # Check if "task_jobs" table is available or not.
-        # Note: A single query with a JOIN is probably a more elegant solution.
-        # However, timing tests suggest that it is cheaper with 2 queries.
-        # This 2nd query may return more results than is necessary, but should
-        # be a very cheap query as it does not have to do a lot of work.
-        if self._db_has_table(
-                self.SUITE_DB, user_name, suite_name, "task_jobs"):
-            stmt = (
-                "SELECT cycle," +
-                " sum(submit_status==1 OR run_status==1) AS n_job_fail" +
-                " FROM task_jobs GROUP BY cycle")
-        else:
-            fail_events_stmt = " OR ".join(
-                ["event=='%s'" % (name) for name in self.STATUSES["fail"]])
-            stmt = (
-                "SELECT cycle," +
-                " sum(" + fail_events_stmt + ") AS n_job_fail" +
-                " FROM task_events GROUP BY cycle")
-        for cycle, n_job_fail in self._db_exec(
-                self.SUITE_DB, user_name, suite_name, stmt):
-            try:
-                entry_of[cycle]["n_states"]["job_fails"] = n_job_fail
-            except KeyError:
-                pass
-
-        return entries, of_n_entries
 
     def get_suite_run_hosts(self, user_name, suite_name, host_names=None):
         """Return a list of host(s) where suite_name has running processes."""
@@ -1009,44 +309,6 @@ class CylcProcessor(SuiteEngineProcessor):
                     fail_list.append(host_name)
         return yes_list, no_list, fail_list
 
-    def get_suite_state_summary(self, user_name, suite_name):
-        """Return a the state summary of a user's suite.
-
-        Return {"is_running": b, "is_failed": b, "server": s}
-        where:
-        * is_running is a boolean to indicate if the suite is running
-        * is_failed: a boolean to indicate if any tasks (submit) failed
-        * server: host:port of server, if available
-
-        """
-        ret = {
-            "is_running": False,
-            "is_failed": False,
-            "server": None}
-        dao = self._db_init(self.SUITE_DB, user_name, suite_name)
-        if not os.access(dao.db_f_name, os.F_OK | os.R_OK):
-            return ret
-
-        port_file_path = os.path.expanduser(
-            os.path.join("~" + user_name, ".cylc", "ports", suite_name))
-        try:
-            port_str, host = open(port_file_path).read().splitlines()
-        except (IOError, ValueError):
-            ret["is_running"] = bool(
-                self._suite_hosts_pgrep(user_name, suite_name)[0])
-        else:
-            ret["is_running"] = True
-            ret["server"] = host.split(".", 1)[0] + ":" + port_str
-
-        stmt = "SELECT status FROM task_states WHERE status GLOB ? LIMIT 1"
-        stmt_args = ["*failed"]
-        for row in self._db_exec(self.SUITE_DB, user_name, suite_name, stmt,
-                                 stmt_args):
-            ret["is_failed"] = True
-            break
-
-        return ret
-
     def get_task_auth(self, suite_name, task_name):
         """
         Return [user@]host for a remote task in a suite.
@@ -1135,11 +397,6 @@ class CylcProcessor(SuiteEngineProcessor):
         """Return Cylc's version."""
         return self.popen("cylc", "--version")[0].strip()
 
-    def is_conf(self, path):
-        """Return "cylc-suite-rc" if path is a Cylc suite.rc file."""
-        if fnmatch(os.path.basename(path), "suite*.rc*"):
-            return "cylc-suite-rc"
-
     def is_suite_registered(self, suite_name):
         """See if a suite is registered
             Return True directory for a suite if it is registered
@@ -1157,9 +414,9 @@ class CylcProcessor(SuiteEngineProcessor):
         cycles = []
         if "*" in items:
             stmt = "SELECT DISTINCT cycle FROM task_events"
-            for row in self._db_exec(self.SUITE_DB, None, suite_name, stmt):
+            for row in self._db_exec(suite_name, stmt):
                 cycles.append(row[0])
-            self._db_close(self.SUITE_DB, None, suite_name)
+            self._db_close(suite_name)
         else:
             for item in items:
                 cycle = self._parse_task_cycle_id(item)[0]
@@ -1169,8 +426,6 @@ class CylcProcessor(SuiteEngineProcessor):
         cwd = os.getcwd()
         self.fs_util.chdir(self.get_suite_dir(suite_name))
         try:
-            stmt = ("UPDATE log_files SET path=?, path_in_tar=? " +
-                    "WHERE cycle==? AND task==? AND submit_num==? AND key==?")
             for cycle in cycles:
                 archive_file_name0 = os.path.join("log",
                                                   "job-" + cycle + ".tar")
@@ -1184,7 +439,7 @@ class CylcProcessor(SuiteEngineProcessor):
                 f_bsize = os.statvfs(".").f_bsize
                 tar = tarfile.open(archive_file_name0, "w", bufsize=f_bsize)
                 for name in names:
-                    cycle, task, s_n, ext = self.parse_job_log_rel_path(name)
+                    cycle, _, s_n, ext = self.parse_job_log_rel_path(name)
                     if s_n == "NN" or ext == "job.status":
                         continue
                     tar.add(name, name.replace("log/", "", 1))
@@ -1194,45 +449,11 @@ class CylcProcessor(SuiteEngineProcessor):
                 self.handle_event(FileSystemEvent(FileSystemEvent.CREATE,
                                                   archive_file_name))
                 self.fs_util.delete(os.path.join("log", "job", cycle))
-                for name in names:
-                    # cycle, task, submit_num, extension
-                    cycle, task, s_n, ext = self.parse_job_log_rel_path(name)
-                    if s_n == "NN" or ext == "job.status":
-                        continue
-                    stmt_args = [
-                        os.path.join(archive_file_name),
-                        name.replace("log/", "", 1),
-                        cycle,
-                        task,
-                        int(s_n),
-                        ext]
-                    self._db_exec(self.JOB_LOGS_DB, None, suite_name,
-                                  stmt, stmt_args, commit=True)
         finally:
             try:
                 self.fs_util.chdir(cwd)
             except OSError:
                 pass
-        self._db_close(self.JOB_LOGS_DB, None, suite_name)
-
-    def job_logs_db_create(self, suite_name, close=False):
-        """Create the job logs database."""
-        if (None, suite_name) in self.daos[self.JOB_LOGS_DB]:
-            return self.daos[self.JOB_LOGS_DB][(None, suite_name)]
-        db_f_name = self.get_suite_dir(suite_name, self.JOB_LOGS_DB)
-        dao = DAO(db_f_name)
-        if os.access(db_f_name, os.R_OK | os.F_OK):
-            return dao
-        self.daos[self.JOB_LOGS_DB][(None, suite_name)] = dao
-        dao.connect(is_new=True)
-        stmt = ("CREATE TABLE log_files(" +
-                "cycle TEXT, task TEXT, submit_num TEXT, key TEXT, " +
-                "path TEXT, path_in_tar TEXT, mtime TEXT, size INTEGER, " +
-                "PRIMARY KEY(cycle, task, submit_num, key))")
-        dao.execute(stmt, commit=True)
-        if close:
-            dao.close()
-        return dao
 
     def job_logs_pull_remote(self, suite_name, items,
                              prune_remote_mode=False, force_mode=False):
@@ -1342,35 +563,6 @@ class CylcProcessor(SuiteEngineProcessor):
         finally:
             self.fs_util.delete(uuid_file_name)
 
-        # Update job log DB
-        dao = self.job_logs_db_create(suite_name)
-        dir_ = self.get_suite_dir(suite_name)
-        stmt = ("REPLACE INTO log_files VALUES(?, ?, ?, ?, ?, ?, ?, ?)")
-        for item in items:
-            cycle, name = self._parse_task_cycle_id(item)
-            if not cycle:
-                cycle = "*"
-            if not name:
-                name = "*"
-            logs_prefix = self.get_suite_dir(
-                suite_name,
-                "log/job/%(cycle)s/%(name)s/*/*" % {
-                    "cycle": cycle, "name": name})
-            for f_name in glob(logs_prefix + "*"):
-                if f_name.endswith("/job.status"):
-                    continue
-                stat = os.stat(f_name)
-                rel_f_name = f_name[len(dir_) + 1:]
-                # cycle, task, submit_num, extension
-                cycle, task, s_n, ext = self.parse_job_log_rel_path(rel_f_name)
-                if s_n == "NN":
-                    continue
-                stmt_args = [cycle, task, int(s_n), ext, rel_f_name, "",
-                             stat.st_mtime, stat.st_size]
-                dao.execute(stmt, stmt_args)
-        dao.commit()
-        dao.close()
-
     def job_logs_remove_on_server(self, suite_name, items):
         """Remove cycle job logs.
 
@@ -1381,9 +573,9 @@ class CylcProcessor(SuiteEngineProcessor):
         cycles = []
         if "*" in items:
             stmt = "SELECT DISTINCT cycle FROM task_events"
-            for row in self._db_exec(self.SUITE_DB, None, suite_name, stmt):
+            for row in self._db_exec(suite_name, stmt):
                 cycles.append(row[0])
-            self._db_close(self.SUITE_DB, None, suite_name)
+            self._db_close(suite_name)
         else:
             for item in items:
                 cycle = self._parse_task_cycle_id(item)[0]
@@ -1415,8 +607,8 @@ class CylcProcessor(SuiteEngineProcessor):
         """Return (cycle, task, submit_num, ext)."""
         return f_name.replace("log/job/", "").split("/", 3)
 
-    @classmethod
-    def process_suite_hook_args(cls, *args, **_):
+    @staticmethod
+    def process_suite_hook_args(*args, **_):
         """Rearrange args for TaskHook.run."""
         task = None
         if len(args) == 3:
@@ -1552,7 +744,7 @@ class CylcProcessor(SuiteEngineProcessor):
             else:
                 ret_code = proc.wait()
                 out, err = proc.communicate()
-                exceptions.append(RosePopenError(keys[2], ret_code, out, err))
+                exceptions.append(RosePopenError(key[2], ret_code, out, err))
         return (sorted(results.values()), exceptions)
 
     def shutdown(self, suite_name, host=None, engine_version=None, args=None,
@@ -1578,36 +770,31 @@ class CylcProcessor(SuiteEngineProcessor):
         self.popen.run_simple(
             *command, env=environ, stderr=stderr, stdout=stdout)
 
-    def _db_close(self, db_name, user_name, suite_name):
+    def _db_close(self, suite_name):
         """Close a named database connection."""
-        key = (user_name, suite_name)
-        if self.daos[db_name].get(key) is not None:
-            self.daos[db_name][key].close()
+        if self.daos.get(suite_name) is not None:
+            self.daos[suite_name].close()
 
-    def _db_exec(self, db_name, user_name, suite_name, stmt, stmt_args=None,
-                 commit=False):
+    def _db_exec(self, suite_name, stmt, stmt_args=None, commit=False):
         """Execute a query on a named database connection."""
-        dao = self._db_init(db_name, user_name, suite_name)
-        return dao.execute(stmt, stmt_args, commit)
+        daos = self._db_init(suite_name)
+        return daos.execute(stmt, stmt_args, commit)
 
-    def _db_has_table(self, db_name, user_name, suite_name, table_name):
+    def _db_has_table(self, suite_name, table_name):
         """Return True if table_name exists in the suite database."""
         cursor = self._db_exec(
-            db_name, user_name, suite_name,
+            suite_name,
             "SELECT name FROM sqlite_master WHERE name==?", [table_name])
         return (cursor and cursor.fetchone() is not None)
 
-    def _db_init(self, db_name, user_name, suite_name):
+    def _db_init(self, suite_name):
         """Initialise a named database connection."""
-        key = (user_name, suite_name)
-        if key not in self.daos[db_name]:
+        if suite_name not in self.daos:
             prefix = "~"
-            if user_name:
-                prefix += user_name
             db_f_name = os.path.expanduser(os.path.join(
-                prefix, self.get_suite_dir_rel(suite_name, db_name)))
-            self.daos[db_name][key] = DAO(db_f_name)
-        return self.daos[db_name][key]
+                prefix, self.get_suite_dir_rel(suite_name, self.SUITE_DB)))
+            self.daos[suite_name] = CylcSuiteDAO(db_f_name)
+        return self.daos[suite_name]
 
     def _parse_task_cycle_id(self, item):
         """Parse name.cycle. Return (cycle, name)."""
@@ -1650,12 +837,12 @@ class CylcProcessor(SuiteEngineProcessor):
         return auth
 
 
-class DAO(object):
+class CylcSuiteDAO(object):
 
     """Generic SQLite Data Access Object."""
 
     CONNECT_RETRY_DELAY = 0.1
-    N_CONNECT_TRIES = 5
+    N_CONNECT_TRIES = 10
 
     def __init__(self, db_f_name):
         self.db_f_name = db_f_name
@@ -1685,7 +872,8 @@ class DAO(object):
             return None
         for _ in range(self.N_CONNECT_TRIES):
             try:
-                self.conn = sqlite3.connect(self.db_f_name)
+                self.conn = sqlite3.connect(
+                    self.db_f_name, self.CONNECT_RETRY_DELAY)
                 self.cursor = self.conn.cursor()
             except sqlite3.OperationalError:
                 sleep(self.CONNECT_RETRY_DELAY)
