+++ t/rose-bush/12-cycle-counts.t
@@ -56,7 +56,8 @@ final cycle point = 20100101T0000Z
 __SUITE_RC__
 export CYLC_CONF_PATH=
 cylc register "${SUITE_NAME}" "${SUITE_DIR}"
-cylc run --debug "${SUITE_NAME}" 2>'/dev/null'
+cylc run --debug "${SUITE_NAME}" 2>'/dev/null' \
+    || cat "${SUITE_DIR}/log/suite/err" >&2
 
 #-------------------------------------------------------------------------------
 # Sort by time_desc
