+++ t/rose-suite-log/06-archive-star.t
@@ -29,7 +29,7 @@ if [[ $TEST_KEY_BASE == *-remote* ]]; then
     JOB_HOST=$(rose host-select -q $JOB_HOST)
 fi
 #-------------------------------------------------------------------------------
-tests 14
+tests 12
 #-------------------------------------------------------------------------------
 # Run the suite.
 export ROSE_CONF_PATH=
@@ -60,23 +60,6 @@ if [[ -n ${JOB_HOST:-} ]]; then
 fi
 set +e
 
-sqlite3 "$HOME/cylc-run/$NAME/log/rose-job-logs.db" \
-    'SELECT path,path_in_tar,key FROM log_files ORDER BY path,path_in_tar ASC;' \
-    >"$TEST_KEY-db-1.out"
-file_cmp "$TEST_KEY-db-1.out" "$TEST_KEY-db-1.out" <<'__OUT__'
-log/job/20130101T0000Z/my_task_1/01/job||job
-log/job/20130101T0000Z/my_task_1/01/job-activity.log||job-activity.log
-log/job/20130101T0000Z/my_task_1/01/job.err||job.err
-log/job/20130101T0000Z/my_task_1/01/job.out||job.out
-log/job/20130101T1200Z/my_task_1/01/job||job
-log/job/20130101T1200Z/my_task_1/01/job-activity.log||job-activity.log
-log/job/20130101T1200Z/my_task_1/01/job.err||job.err
-log/job/20130101T1200Z/my_task_1/01/job.out||job.out
-log/job/20130102T0000Z/my_task_1/01/job||job
-log/job/20130102T0000Z/my_task_1/01/job-activity.log||job-activity.log
-log/job/20130102T0000Z/my_task_1/01/job.err||job.err
-log/job/20130102T0000Z/my_task_1/01/job.out||job.out
-__OUT__
 N_JOB_LOGS=$(wc -l "$TEST_KEY-list-job-logs-before.out" | cut -d' ' -f1)
 run_pass "$TEST_KEY-command" rose suite-log -n $NAME --archive '*' --debug
 run_fail "$TEST_KEY-list-job-logs-after" ls $SUITE_RUN_DIR/log/job/*
@@ -109,35 +92,6 @@ N_JOB_LOGS_ARCH=$(echo "$ALL_JOB_LOGS_ARCH" | wc -l | cut -d' ' -f1)
 run_pass "$TEST_KEY-n-arch" test "${N_JOB_LOGS}" -eq "${N_JOB_LOGS_ARCH}"
 
 file_cmp "$TEST_KEY-command.err" "$TEST_KEY-command.err" </dev/null
-sqlite3 "$HOME/cylc-run/$NAME/log/rose-job-logs.db" \
-    'SELECT path,path_in_tar,key FROM log_files ORDER BY path,path_in_tar ASC;' \
-    >"$TEST_KEY-db-2.out"
-file_cmp "$TEST_KEY-db-2.out" "$TEST_KEY-db-2.out" <<'__OUT__'
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_1/01/job|job
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_1/01/job-activity.log|job-activity.log
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_1/01/job.err|job.err
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_1/01/job.out|job.out
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_2/01/job|job
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_2/01/job-activity.log|job-activity.log
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_2/01/job.err|job.err
-log/job-20130101T0000Z.tar.gz|job/20130101T0000Z/my_task_2/01/job.out|job.out
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_1/01/job|job
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_1/01/job-activity.log|job-activity.log
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_1/01/job.err|job.err
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_1/01/job.out|job.out
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_2/01/job|job
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_2/01/job-activity.log|job-activity.log
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_2/01/job.err|job.err
-log/job-20130101T1200Z.tar.gz|job/20130101T1200Z/my_task_2/01/job.out|job.out
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_1/01/job|job
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_1/01/job-activity.log|job-activity.log
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_1/01/job.err|job.err
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_1/01/job.out|job.out
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_2/01/job|job
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_2/01/job-activity.log|job-activity.log
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_2/01/job.err|job.err
-log/job-20130102T0000Z.tar.gz|job/20130102T0000Z/my_task_2/01/job.out|job.out
-__OUT__
 #-------------------------------------------------------------------------------
 rose suite-clean -q -y $NAME
 exit 0
